requirements.txt
streamlit>=1.44,<2
pandas>=2.2,<3
numpy>=2.1,<3
plotly>=6.0,<7
networkx>=3.4,<4
requests>=2.32,<3
python-dotenv>=1.0,<2
urllib3>=2.3,<3
